Pubsub service setup
