+++ template/node/Cargo.toml
@@ -17,6 +17,7 @@ futures = "0.3.4"
 log = "0.4.8"
 structopt = "0.3.8"
 jsonrpc-core = "14.0.3"
+jsonrpc-pubsub = "14.0.5"
 
 sp-api = { version = "2.0.0-dev", path = "../../vendor/substrate/primitives/api" }
 sp-blockchain = { version = "2.0.0-dev", path = "../../vendor/substrate/primitives/blockchain" }
