+++ template/node/src/service.rs
@@ -154,6 +154,7 @@ pub fn new_full(config: Configuration, manual_seal: bool) -> Result<TaskManager,
 	let prometheus_registry = config.prometheus_registry().cloned();
 	let telemetry_connection_sinks = sc_service::TelemetryConnectionSinks::default();
 	let is_authority = role.is_authority();
+	let subscription_task_executor = sc_rpc::SubscriptionTaskExecutor::new(task_manager.spawn_handle());
 
 	let rpc_extensions_builder = {
 		let client = client.clone();
@@ -168,7 +169,10 @@ pub fn new_full(config: Configuration, manual_seal: bool) -> Result<TaskManager,
 				is_authority,
 				command_sink: Some(command_sink.clone())
 			};
-			crate::rpc::create_full(deps)
+			crate::rpc::create_full(
+				deps, 
+				subscription_task_executor.clone()
+			)
 		})
 	};
